Month 14/step 2

+++ src/controllers/cards.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from 'express';
 import Errors from '../errors/errors';
 import Card from '../models/card';
-import { AuthRequest } from '../middlewares/authorization';
+import { AuthRequest } from '../middlewares/auth';
 
 export const getCards = (req: Request, res: Response, next: NextFunction) => {
   Card.find({})
@@ -35,14 +35,22 @@ export const postCard = (req: AuthRequest, res: Response, next: NextFunction) =>
     });
 };
 
-export const deleteCard = (req: Request, res: Response, next: NextFunction) => {
-  Card.findByIdAndRemove(req.params.cardId)
+export const deleteCard = (req: AuthRequest, res: Response, next: NextFunction) => {
+  const userId = req.user?._id;
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (!card) {
         throw Errors.notFoundRequest();
       }
+      if (card.owner.toString() !== userId) {
+        throw Errors.forbiddenError();
+      }
+      return Card.findByIdAndRemove(req.params.cardId);
+    })
+    .then((card) => {
       res.send({
         message: 'Карточка удалена',
+        data: card,
       });
     })
     .catch((err) => {
